Convert Answers component to TS
